Core/Players: fixed structure of CMSG_SET_ACTION_BUTTON and downgraded player actions
https://github.com/TrinityCore/TrinityCore/commit/9dd8ced81c21af34c3b578b66768f04701b23a44

+++ Source/Framework/Constants/PlayerConst.cs
@@ -316,7 +316,7 @@ namespace Framework.Constants
         FFAPvP = 0x40
     }
 
-    public enum ActionButtonUpdateState : uint
+    public enum ActionButtonUpdateState : byte
     {
         UnChanged = 0,
         Changed = 1,
@@ -324,7 +324,7 @@ namespace Framework.Constants
         Deleted = 3
     }
 
-    public enum ActionButtonType : uint
+    public enum ActionButtonType : byte
     {
         Spell = 0x00,
         C = 0x01,                         // click?
